requirements.txt
pip
requests
flask<0.12
matplotlib
Pillow
pytest
